post authorization update

+++ backend/routes/posts.js
@@ -26,12 +26,13 @@ const storage = multer.diskStorage({
     }
 });
 
-router.post('/api/post', multer({ storage: storage }).single('image'), (req, res, next) => {
+router.post('/api/post', checkAuth, multer({ storage: storage }).single('image'), (req, res, next) => {
     const url = req.protocol + '://' + req.get('host');
     const post = new PostModel({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename
+        imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId
     });
     post.save()
         .then(createdPost => {
@@ -57,13 +58,20 @@ router.put('/api/post/:id',
         _id: req.params.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
-    PostModel.updateOne({_id: req.params.id}, updatedPost)
+    PostModel.updateOne({_id: req.params.id, creator: req.userData.userId}, updatedPost)
         .then(result => {
-            res.status(200).json({
-                message: 'Update successul!'
-            })
+            if(result.modifiedCount > 0) {
+                res.status(200).json({
+                    message: 'Update successful!'
+                });
+            } else {
+                res.status(401).json({
+                    message: 'Not authorized!'
+                });
+            }
         });
 });
 
@@ -102,9 +110,12 @@ router.get('/api/posts', (req, res, next) => {
 });
 
 router.delete('/api/post/:id', checkAuth, (req, res, next) => {
-    PostModel.deleteOne({_id: req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({message: 'Post deleted'});
+    PostModel.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+        if(result.deletedCount > 0) {
+            res.status(200).json({message: 'Post deleted'});
+        } else {
+            res.status(401).json({message: 'Not authorized!'});
+        }
     });
 });
